+++ lib/yust.dart
@@ -5,6 +5,7 @@ library yust;
 
 export 'src/extensions/date_time_extension.dart';
 export 'src/extensions/list_extension.dart';
+export 'src/extensions/server_now.dart';
 export 'src/extensions/string_extension.dart';
 export 'src/models/yust_doc.dart';
 export 'src/models/yust_doc_setup.dart';
@@ -18,6 +19,11 @@ export 'src/services/yust_database_service_mocked.dart';
 export 'src/util/google_cloud_helpers.dart';
 export 'src/util/google_cloud_helpers_shared.dart';
 export 'src/util/object_helper.dart';
+export 'src/util/yust_database_statistics.dart'
+    show
+        YustStatisticsMap,
+        YustAggregatedStatisticsMap,
+        YustEnhancedStatisticsMap;
 export 'src/util/yust_exception.dart';
 export 'src/util/yust_field_transform.dart';
 export 'src/util/yust_helpers.dart';